backend/requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
